+++ lev/agents/factory.py
@@ -1,7 +1,7 @@
 from lev.agents.reasoning import ReasoningAgent
 from lev.agents.tool import ToolsAgent
 from lev.config import Eval
-from lev.controller import Introspector
+from lev.workflow import Introspector
 from lev.core.agent import SimpleAgent
 from lev.core.llm_provider import LlmProvider
 from lev.loader import get_persona_system_prompt
